requirements.txt
httpx>=0.27
python-dotenv>=1.0
playwright>=1.40
authlib>=1.3
